扩展websocket

+++ src/main/java/com/example/security/LogoutSuccessHandlerImpl.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
 import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -26,7 +27,9 @@ public class LogoutSuccessHandlerImpl implements LogoutSuccessHandler {
     @Override
     public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
         String token = TokenFilter.getToken(request);
-        tokenService.deleteToken(token);
+        if (!StringUtils.isEmpty(token) && !"null".equals(token)) {
+            tokenService.deleteToken(token);
+        }
         ResultVO resultVO = new ResultVO<>(SUCCESS, "退出成功！", null);
         ResponseUtil.println(response, resultVO);
     }
